bugfix: non-active repeated 1st char in ending delim
listings.dtx v1.4 fails to produce correct output for the document
below. The reason is that > is made active while scanning for the
ending delimiter, but the second > in >> is not an active character.
Therefore, v1.4 does not correctly recognize an existing >> in the
input line.
\documentclass{article}
\usepackage{listings}
\begin{document}
\begin{lstlisting}[escapeinside={<<}{>>}]
a <<aaa>> 77
b <<bbb:$\int_\Omega$>> end
c ccc
\end{lstlisting}
\end{document}